Declarative formats and initial fl and out-string
A very basic Format-List is now supported, you can specify property
names. The directives are attached to the objects and the original
objects are written to the pipe, so you can do things like:

    $processes = ps | fl Name,Id
    $processes | Where Name -like p* | Out-PSMore

A basic Out-String is also added.

Initial support for a declarative formatting is now supported via
attributes. The `DefaultDisplayProperty` attribute replaces the need for
the special member in the type table.

The `FormatProxy` attribute specifies that the type is a proxy for
another type likely not defined in the same assembly.

Future work will add support for named views so you can specify
different formats with a single type definition.

Initial performance looks very good, about an order of magnitude faster:

    Get-Process | Format-List Name,Id | Out-String -Stream

This runs in ~170ms whereas

    Get-Process | Format-PSMoreList Name,Id | Out-PSMoreString -Stream

Runs in ~50ms the first time, and ~20ms after that.